+++ database/migrations/2024_01_30_134158_create_orders_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->datetime('order_date');
             $table->integer('order_number');
             $table->integer('customer_id')->nullable(true);
-            $table->float('total_amount')->default(0);
+            $table->integer('total_amount')->default(0);
             $table->integer('id_cell')->nullable(true);
         });
     }
